split key.rs into two smaller files
