+++ grug/storage/src/lib.rs
@@ -3,12 +3,14 @@ mod codec;
 mod counter;
 mod index;
 mod item;
-mod key;
 mod map;
 mod path;
 mod prefix;
+mod prefixer;
+mod primary_key;
 mod set;
 
 pub use {
-    bound::*, codec::*, counter::*, index::*, item::*, key::*, map::*, path::*, prefix::*, set::*,
+    bound::*, codec::*, counter::*, index::*, item::*, map::*, path::*, prefix::*, prefixer::*,
+    primary_key::*, set::*,
 };
